+++ src/test/scala/io/iohk/ethereum/consensus/pow/MinerSpecSetup.scala
@@ -53,10 +53,10 @@ trait MinerSpecSetup extends ConsensusConfigBuilder with MockFactory {
       value = BigInt("108516826677274384"),
       payload = ByteString.empty
     ),
-    pointSign = 0x9d.toByte,
+    pointSign = 0x9d.toInt,
     signatureRandom = ByteString(Hex.decode("beb8226bdb90216ca29967871a6663b56bdd7b86cf3788796b52fd1ea3606698")),
     signature = ByteString(Hex.decode("2446994156bc1780cb5806e730b171b38307d5de5b9b0d9ad1f9de82e00316b5")),
-    chainId = 0x3d.toByte
+    chainId = 0x3d.toInt
   )
 
   lazy val consensus: PoWConsensus = buildPoWConsensus().withBlockGenerator(blockGenerator)
